A few flutter things
